move fixed file names
